my_app and timeloop: first attempt at an iteration 0 startup
This is the first attempt at an initialization procedure that is more
efficient than what is currently done. This commit starts to implement a
checkpoint during a run() on the coarsest level. Some issues:
1) The checkpoint only stores one time point, we need to store a
different checkpoint for each c-point.
2) After this app object is called, there is a segfault. Probably some
parameteracceptor issue. This is less than ideal.

Does this mean we ought to do something different, using only the app we
have, and the px from a single processor? How do we ensure that this is
doing what we want on each time processor.

We could always run two separate programs to avoid segfault, making a
.sh file with approprate calls to each program. This could be ok, but
timing might be strange.