adds logic to time_loop to make sure checkpointing is exactly at C-points
This change is with a bool use_cycle_in_name_ in the time_loop class,
which is default false and can be changed by calling
set_use_cycle_in_name(), and should not affect functionality. When
toggled on, it shortens some time steps so that we write checkpoint
files only at the "exact" C-points.

Otherwise, this "exactness" on checkpointing is turned off. Note that
this is different that the "enforce_t_final_" which only affects the
very last time step length.