my_app.template.h: bugfix where this_mpi_process() crashed
This bug is due to the fact that the ensemble was trying to do global t
synchronization inside the step() function, and there probably was a
clash in MPI_COMM_WORLD where another processor was doing something that
prevented this_mpi_process() from working.

The fix is to ensure that the mpi_ensemble only synchronizes over the
"local" communicators. Hence setting global_synchronization=false in the
ensemble's constructor.