my_app.template.h: bugfix where r=0 and a user refinement r>0 caused a segfault
if the user defined coarsest level r>0, then a segfault happened because
unrefined_level assumes r=0.

The solution was to reinitialize and prepare the associated data
structures by adding a call to offline_data.prepare(...) after the
triangulation was loaded into the triangulation.

This initialized the data structures in unrefined_level with the new
fact that r>0 is true. We still reset offline_data at the end, as we
assume that the Init() function is invariant with respect to
unrefined_level.