Init(): t=0 needs to be initialized on the fine level, not read from coarse checkpoint.
This gets rid of a bug where the initial condition on brick 0 looked
like the initial condition on brick 0 from the coarse level. We want to
use the one from the fine level, so add a special case to Init() which
skips the load for c_id==0 and instead interpolates this brick's initial
condition from the initial_values on the finest level.