mgrit_function::enforce_physicality WIP on a local stencil refactor
Instead of looping over all global cells, we enforce physicality
constraings on local stencils, taking care to modify a copy so as not to
double limit as we proceed over local nodes. This is WIP since the
behavior is not quite right as can be seen in the test in
project_to_reasonable_bounds.cc's output.